+++ CHANGELOG.md
@@ -9,7 +9,7 @@ All notable changes to this project will be documented in this file.
 - New `query` plugin. Learn more about it in the [documentation](https://radonis.vercel.app/docs/plugins/query).
 
 ### Changed
-- Requests to an endpoint rendering a view will now return the views props as JSON when the `accept=application/json` header is present in the request.
+- Requests to an endpoint rendering a view will now return the view's props as JSON when the `accept=application/json` header is present in the request.
 - Updated dependencies.
 
 ## [1.8.9] - 2022-09-10